sign-skeleton-pipeline/requirements.txt
numpy
opencv-python
torch
torchvision
scikit-learn
matplotlib
pandas
tqdm
imageio
imageio[ffmpeg]
pyyaml
requests